refactor(super_diag_tensor): Moved 'super_diag_tensor' into generation utils.
