+++ hottbox/core/tests/test_structures.py
@@ -3,8 +3,8 @@ import sys
 import io
 import numpy as np
 from functools import reduce
-from ..structures import *
-from ..operations import unfold, kolda_unfold
+from ..structures import Tensor, BaseTensorTD, TensorCPD, TensorTKD, TensorTT, residual_tensor
+from ..operations import unfold, kolda_unfold, mode_n_product
 from .._meta import State
 from ...errors import TensorModeError, TensorShapeError, TensorStateError, TensorTopologyError, ModeError, StateError
 
@@ -2143,64 +2143,6 @@ class TestTensorTT:
         tensor_tt.reset_mode_index(mode=0)
         tensor_tt_2 = TensorTT(core_values=core_values)
         assert all([tensor_tt.modes[i].index == tensor_tt_2.modes[i].index for i in range(tensor_tt.order)])
-
-
-def test_super_diag_tensor():
-    """ Tests for creating super-diagonal tensor"""
-    order = 3
-    rank = 2
-    correct_shape = (rank, ) * order
-    true_default_data = np.array([[[1., 0.],
-                                   [0., 0.]],
-
-                                  [[0., 0.],
-                                   [0., 1.]]])
-    true_default_mode_names = ['mode-0', 'mode-1', 'mode-2']
-    correct_values = np.arange(rank)
-    true_data = np.array([[[0., 0.],
-                           [0., 0.]],
-
-                          [[0., 0.],
-                           [0., 1.]]])
-
-    # ------ tests for default super diagonal tensor
-    tensor = super_diag_tensor(correct_shape)
-    assert isinstance(tensor, Tensor)
-    np.testing.assert_array_equal(tensor.data, true_default_data)
-    assert (tensor.mode_names == true_default_mode_names)
-
-    # ------ tests for super diagonal tensor with custom values on the main diagonal
-    tensor = super_diag_tensor(correct_shape, values=correct_values)
-    assert isinstance(tensor, Tensor)
-    np.testing.assert_array_equal(tensor.data, true_data)
-    assert (tensor.mode_names == true_default_mode_names)
-
-    # ------ tests that should Fail
-
-    with pytest.raises(TypeError):
-        # shape should be passed as tuple
-        super_diag_tensor(shape=list(correct_shape))
-
-    with pytest.raises(ValueError):
-        # all values in shape should be the same
-        incorrect_shape = [rank] * order
-        incorrect_shape[1] = order+1
-        super_diag_tensor(shape=tuple(incorrect_shape))
-
-    with pytest.raises(ValueError):
-        # values should be an one dimensional numpy array
-        incorrect_values = np.ones([rank, rank])
-        super_diag_tensor(shape=correct_shape, values=incorrect_values)
-
-    with pytest.raises(ValueError):
-        # too many values for the specified shape
-        incorrect_values = np.ones(correct_shape[0]+1)
-        super_diag_tensor(shape=correct_shape, values=incorrect_values)
-
-    with pytest.raises(TypeError):
-        # values should be a numpy array
-        incorrect_values = [1] * correct_shape[0]
-        super_diag_tensor(shape=correct_shape, values=incorrect_values)
 
 
 def test_residual_tensor():
